fix(webpack): correctly throw compilation errors (#1027)
This PR corrects #863 where if a syntax error is encountered in tests,
the process throws an error during compilation. Tests are never run and
CI builds should now fail if there are syntax errors. This is done by
adding a small webpack plugin that checks for errors during compilation,
and adds them to an array of errors. This array is then thrown but only
if there are errors during the compilation stage.

Based on the suggestions from:
https://gist.github.com/Stuk/6b574049435df532e905